01_docker_terraform/pyproject.toml
[project]
name = "01-docker-terraform"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "pandas>=3.0.0",
    "psycopg2-binary>=2.9.11",
    "psycopg[binary,pool]>=3.3.2",
    "pyarrow>=23.0.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.46",
]
